refactor properties files to dedicated OutputMetaData class
Properties files are much less memory efficient than a direct class
and are a major cause of memory pressure / garbage collection.
This is also a step towards enabling a different storage mechanism
than properties files for this data. For now however the storage
is still properties files, but at least the serialization format
is now isolated from the rest of Bpipe.