add 'allowForeign' option to 'from' to let it process non-outputs
Recently a change was added to strictly check that bpipe only references
files actually specified as inputs or created as outputs. However in
some pipelines this check can fail if the pipeline was relying on
recognising files that were not declared as outputs. This can be
rectified now by adding allowForeign as an option:

from('*.bam', allowForeign:true) { ... }