do not terminate parent branch when child succeeds
currently this limits the use of succeed which otherwise could be very
useful to execute local control flow. since it terminates parent branches
all the way to root, currently this aborts whole pipeline even if the
temrination is only local. This constitutes a behaviour change so needs
evaluation for impact.