initial support for formal pipeline channels : resumable branches linked to a variable name
we find a very common pattern occuring where pipeline lines are split
long lines that occur repeatedly within the pipeline and which often
utilise the same outputs from previous lines. For example, when
processing many samples in parallel. Channels are like branches
but when used repeatedly, automatically link inputs between
otherwise disjoint parts of the pipeline based on the
branch name used. Channels are also linked to a named
"branch variable" that identifies the active branch to running
commands and is a constant throughout each strand within a parallel
section.