change container support to ignore default entry point
this is needed because the default entry point for most containers
is designed to emulate the command they are provisioning, but this
does not easily allow portable commands that look identical when
they are run inside or outside a container.

previous behaviour can be restored for docker by setting
entryPoint="default" in bpipe.config for the container config.