support for hooks and durable JMS agent replies
because: agent reply functionality relied on the agent owning
the process that executed bpipe. If the pipeline was retried
manually, the agent would never send the reply, making it hard to
implement robust orchestrated workflows using the Bpipe agent.

Note that using the hook functionality to send agent replies is
now the default, but old behaviour can be restored by setting
agentMode="direct" in the agent bpipe.config